Moved stack to construct
