+++ src/release_artifacts_resources/ios/cdk/cdk/credential_rotation_stack.py
@@ -1,0 +1,8 @@
+from aws_cdk import core as cdk
+
+
+class CredentialRotationStack(cdk.Stack):
+    def __init__(self, scope: cdk.Construct, construct_id: str, **kwargs) -> None:
+        super().__init__(scope, construct_id, **kwargs)
+
+        # TODO: Incomplete implementation
